# lib/kernel/http/openapi-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "yetkin.ai API v1",
    "version": "1",
    "description": "yetkin.ai Modüler Monolit + API-First Dron sözleşmesi — Proof / Payments. Faz 1 kamu sözleşmesi B2C'dir; Marketplace tag'i ve freelancer path'leri yayınlanmaz (iç bounded-context adı durur). Kopya `app/api/v1` handler ağacı yoktur; kenar soyar. Zarf: { ok, error, requestId, apiVersion, data }. X-Rail-Min-Version API sözleşme sürümüdür; mağaza uygulama build'i değildir (X-Rail-App-Build ayrı, v1'de kapı değil). Amiral ve Dron aynı v1 zarfı konuşur. Versiyonsuz `{ ok, ...data }` serimi kapalıdır. Yayınlanmış alan sessizce düşmez. CORS joker ve Allow-Credentials yoktur. SHA-256 akademi kaydı kriptografik imza değildir; içerik özeti / bütünlük kaydıdır."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/wallet/top-up": {
      "post": {
        "operationId": "wallet-top-up",
        "tags": [
          "Payments"
        ],
        "summary": "/api/v1/wallet/top-up",
        "description": "Kanonik yol: /api/wallet/top-up. Auth: Bearer JWT. Çerez oturumu yok sayılır (cookieAuth=false). Idempotency-Key UUID zorunlu (Idempotency-Key). Yayınlanmış data kökleri: merchantOid, token, iframeUrl, sandboxMode, alreadySettled, mockCheckout, status, checkoutPassportUrl.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-rail-min-version",
            "in": "header",
            "required": true,
            "description": "Asgari API sürümü. Health dışında zorunlu. Eski istemci 426 alır.",
            "schema": {
              "type": "string",
              "pattern": "^[1-9]\\d*$"
            }
          },
          {
            "name": "x-request-id",
            "in": "header",
            "required": false,
            "description": "İsteğe bağlı korelasyon UUID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer JWT. v1 çerez oturumunu yok sayar.",
            "schema": {
              "type": "string",
              "pattern": "^Bearer\\s+\\S+"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "UUID. Aynı anahtar ikinci debit doğurmaz.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "v1 başarı zarfı. Yayınlanmış data alanları sessizce düşmez.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "error": {
                      "type": "string",
                      "nullable": true,
                      "enum": [
                        null
                      ]
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "apiVersion": {
                      "type": "string",
                      "enum": [
                        "1"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "merchantOid": {
                          "type": "string",
                          "minLength": 1
                        },
                        "token": {
                          "nullable": true,
                          "type": "string",
                          "minLength": 1
                        },
                        "iframeUrl": {
                          "nullable": true,
                          "type": "string",
                          "minLength": 1
                        },
                        "sandboxMode": {
                          "type": "boolean"
                        },
                        "alreadySettled": {
                          "type": "boolean"
                        },
                        "mockCheckout": {
                          "type": "boolean"
                        },
                        "status": {
                          "nullable": true,
                          "type": "string",
                          "enum": [
                            "PENDING",
                            "PAID",
                            "CLEARED",
                            "FAILED"
                          ]
                        },
                        "checkoutPassportUrl": {
                          "nullable": true,
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "merchantOid",
                        "token",
                        "iframeUrl",
                        "sandboxMode",
                        "alreadySettled",
                        "mockCheckout",
                        "status",
                        "checkoutPassportUrl"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "ok",
                    "error",
                    "requestId",
                    "apiVersion",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Sürüm başlığı / doğrulama / Idempotency-Key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RailV1FailEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Oturum gerekli.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RailV1FailEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency-Key aynı anahtarla farklı gövde kullanılamaz.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RailV1FailEnvelope"
                }
              }
            }
          },
          "426": {
            "description": "Bu uygulama güncel değil. yetkin.ai uygulamasını mağazadan güncelle.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RailV1FailEnvelope"
                }
              }
            }
          }
        },
        "x-rail-cookie-auth": false,
        "x-rail-idempotency": true,
        "x-rail-route-auth": "session",
        "x-rail-published-data-paths": [
          "merchantOid",
          "token",
          "iframeUrl",
          "sandboxMode",
          "alreadySettled",
          "mockCheckout",
          "status",
          "checkoutPassportUrl"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amountMinor": {
                    "type": "integer",
                    "minimum": 1000,
                    "maximum": 2000000
                  },
                  "distanceContractAccepted": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  },
                  "digitalImmediatePerformanceAccepted": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  },
                  "consentVersion": {
                    "type": "string",
                    "enum": [
                      "2026-09-05"
                    ]
                  },
                  "billing": {
                    "type": "object",
                    "properties": {
                      "invoiceType": {
                        "type": "string",
                        "enum": [
                          "individual",
                          "corporate"
                        ]
                      },
                      "fullName": {
                        "type": "string"
                      },
                      "tckn": {
                        "nullable": true,
                        "type": "string"
                      },
                      "companyTitle": {
                        "type": "string"
                      },
                      "taxOffice": {
                        "type": "string"
                      },
                      "vkn": {
                        "type": "string"
                      },
                      "phone": {
                        "type": "string"
                      },
                      "address": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "invoiceType"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "amountMinor",
                  "distanceContractAccepted",
                  "digitalImmediatePerformanceAccepted",
                  "consentVersion",
                  "billing"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase vatandaş JWT. Cookie güvenlik şeması yoktur."
      }
    },
    "schemas": {
      "RailV1FailEnvelope": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string",
            "minLength": 1
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "apiVersion": {
            "type": "string",
            "enum": [
              "1"
            ]
          },
          "data": {
            "type": "string",
            "nullable": true,
            "enum": [
              null
            ]
          }
        },
        "required": [
          "ok",
          "error",
          "requestId",
          "apiVersion",
          "data"
        ],
        "additionalProperties": false
      },
...
    },
...
  }
}